Use a custom `LiveView` to fix double pre rendering
